refactor: 공연 이벤트를 예매 이벤트와 분리
